+++ yash-syntax/src/parser/lex/keyword.rs
@@ -44,6 +44,21 @@ pub enum Keyword {
     CloseBrace,
 }
 
+impl Keyword {
+    /// Determines if this token can be a delimiter of a clause.
+    ///
+    /// This function returns `true` for `Do`, `Done`, `Elif`, `Else`, `Esac`,
+    /// `Fi`, `Then`, and `CloseBrace`, and `false` for others.
+    pub fn is_clause_delimiter(self) -> bool {
+        use Keyword::*;
+        match self {
+            Do | Done | Elif | Else | Esac | Fi | Then | CloseBrace => true,
+            Bang | OpenBracketBracket | Case | For | Function | If | In | Until | While
+            | OpenBrace => false,
+        }
+    }
+}
+
 impl fmt::Display for Keyword {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         use Keyword::*;